Updated README
Tested if the background download really survives app termination on android (it does)
Added Qurik for iOS: when the server is offline, iOS will try to download later. When exactly "later" is remains a mystery, the iOS default settings allow this be up to one week later. Maybe that could need some tweaking...